mise en place de la function update des devises

+++ database/migrations/2023_07_18_182401_create_paires_table.php
@@ -17,7 +17,7 @@ class CreatePairesTable extends Migration
             $table->id();
             $table->string('devise_1', 3);
             $table->string('devise_2', 3);
-            $table->float('taux',8,6);
+            $table->integer('taux');
             $table->unsignedInteger('conversionNumber')->default(0);
             $table->timestamps();
         });
